transcribe/requirements.txt
openai-whisper==20250625
ffmpeg-python==0.2.0
torch==2.2.2
numpy==1.26.4
numba==0.62.0
llvmlite==0.45.1
tiktoken==0.12.0
tqdm==4.67.3
